Cleanup pre-workcell editor migration
