+++ rmf_site_editor/src/workcell/load.rs
@@ -30,8 +30,8 @@ use bevy::prelude::*;
 use std::collections::HashSet;
 
 use rmf_site_format::{
-    Category, ConstraintDependents, FrameMarker, Geometry, MeshConstraint, ModelMarker,
-    NameInWorkcell, Parented, Scale, SiteID, Workcell, WorkcellModel,
+    Category, FrameMarker, Geometry, ModelMarker, NameInWorkcell, Parented, Scale, SiteID,
+    Workcell, WorkcellModel,
 };
 
 #[derive(Event)]
@@ -53,8 +53,6 @@ fn generate_workcell_entities(
     let mut id_to_entity = HashMap::new();
     // Hashmap of parent id to list of its children entities
     let mut parent_to_child_entities = HashMap::new();
-    // Hashmap of parent model entity to constraint dependent entity
-    let mut model_to_constraint_dependent_entities = HashMap::new();
 
     let root = commands
         .spawn(SpatialBundle::INHERITED_IDENTITY)
@@ -80,7 +78,6 @@ fn generate_workcell_entities(
                 }
                 Geometry::Mesh { source, scale } => {
                     commands.entity(e).insert((
-                        ConstraintDependents::default(),
                         NameInWorkcell(parented.bundle.name.clone()),
                         parented.bundle.pose.clone(),
                         Scale(scale.unwrap_or(Vec3::ONE)),
@@ -114,21 +111,6 @@ fn generate_workcell_entities(
             .insert(SiteID(*id))
             .insert(FrameMarker)
             .id();
-        if let Some(c) = &parented_anchor.bundle.mesh_constraint {
-            let model_entity = *id_to_entity
-                .get(&c.entity)
-                .expect("Mesh constraint refers to non existing model");
-            commands.entity(e).insert(MeshConstraint {
-                entity: model_entity,
-                element: c.element.clone(),
-                relative_pose: c.relative_pose,
-            });
-            let constraint_dependents: &mut HashSet<Entity> =
-                model_to_constraint_dependent_entities
-                    .entry(model_entity)
-                    .or_default();
-            constraint_dependents.insert(e);
-        }
         if let Some(name) = &parented_anchor.bundle.name {
             commands.entity(e).insert(name.clone());
         }
@@ -163,13 +145,6 @@ fn generate_workcell_entities(
             .or_default();
         child_entities.push(e);
         id_to_entity.insert(*id, e);
-    }
-
-    // Add constraint dependents to models
-    for (model, dependents) in model_to_constraint_dependent_entities {
-        commands
-            .entity(model)
-            .insert(ConstraintDependents(dependents));
     }
 
     for (parent, children) in parent_to_child_entities {
