+++ rmf_site_format/src/lib.rs
@@ -98,6 +98,9 @@ pub use physical_camera::*;
 pub mod point;
 pub use point::*;
 
+pub mod primitive_shape;
+pub use primitive_shape::*;
+
 pub mod recall;
 pub use recall::*;
